third_party/WebKit/LayoutTests/fast/encoding/api/end-of-file.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>

description("Edge cases around non-fatal errors at EOF");

shouldThrow("new TextDecoder('utf-8', {fatal: true}).decode(new Uint8Array([0xff]))");

debug("");
debug("Should not throw or hang:");
evalAndLog("new TextDecoder('utf-8').decode(new Uint8Array([0xff]))");

debug("");
shouldThrow("new TextDecoder('utf-16le', {fatal: true}).decode(new Uint8Array([0x00]))");

debug("");
debug("Should not throw or hang:");
evalAndLog("new TextDecoder('utf-16le').decode(new Uint8Array([0x00]))");

debug("");
shouldThrow("new TextDecoder('utf-16be', {fatal: true}).decode(new Uint8Array([0x00]))");

debug("");
debug("Should not throw or hang:");
evalAndLog("new TextDecoder('utf-16be').decode(new Uint8Array([0x00]));");

</script>
<script src="../../js/resources/js-test-post.js"></script>
